Reduce default radius by half
The default setting of 1000m was showing FAR more pokestops than what the game shows. By doing some trial and error in an area with a lot of pokestops, I found that using 500m resulted in an accurate number of pokestops when compared to what I see in-game. I guess it makes sense that the diameter is exactly 1 km.